Fix URLs and copy
The 2015 site is being moved from `https://pygotham.org/2015/` to
`https://2015.pygotham.org/`. The URLs for all 2015 pages are being
fixed to reflect this.

At the same time, some of the copy (for both 2014 and 2015) is being
updated to reflect the fact that the sites are now archived. This also
includes updating outdated email addresses.